+++ frontend/src/Components/ShrimpFishing/FriedShrimpIcon.tsx
@@ -1,0 +1,53 @@
+export function FriedShrimpIcon() {
+  return (
+    <svg
+      width="512"
+      height="512"
+      viewBox="0 0 512 512"
+      style={{
+        height: 'auto',
+      }}
+      xmlns="http://www.w3.org/2000/svg"
+      className="h-full w-full"
+    >
+      <rect
+        width="512"
+        height="512"
+        x="0"
+        y="0"
+        rx="30"
+        fill="transparent"
+        stroke="transparent"
+        strokeWidth="0"
+        strokeOpacity="100%"
+        paintOrder="stroke"
+      ></rect>
+      <svg
+        width="256px"
+        height="256px"
+        viewBox="0 0 64 64"
+        fill="currentColor"
+        x="128"
+        y="128"
+        role="img"
+        // style="display:inline-block;vertical-align:middle"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <g fill="currentColor">
+          <path
+            fill="#f46c6c"
+            d="M27.7 12.2c-3.8 5.2-23.1 9.6-23.1 9.6S7.4 16.4 15 13c0 0-8 3.3-13 1.5c0 0 7.7-8.7 24-10.2l1.7 7.9"
+          />
+          <path
+            fill="#f8d317"
+            d="M60.9 27.1c-.7-2.8.6-5.8-1.7-6.9c-1.1-.5-1.8-2.1-2-3.3c-.3-1.3-.1-2.3-1.1-3.2c-1-.9-2.7-1.5-3.7-2.4c-.9-.9-.6-2.2-1.3-3.2c-.9-1.3-2.1-1.7-3.6-1.9c-1.4-.2-2.5-.7-3.6-1.6c-1.1-.9-2.2-1.9-3.7-1.8c-1.2 0-2.4.6-3.6.5c-1.4 0-2.3-1.3-3.7-1.3c-1.1.1-1.9 1.1-2.8 1.5c-1.1.5-2.2-.1-3.4 0c-1.1.1-1.7 1.1-2.2 2c-.7 1.3-1.9 1.4-3 2.3c-.7.5-.9 1.7 0 3.3c.8 1.3.6 3.2 1.7 4.4c1.1 1.2 3.5-.6 4.8.1c1.6.8 2.1 1.7 4 2c1.2.2 2.9 2.7 3.7 3.2c1.2.6 3 1.8 3.7 2.8c.6.9 1.1 3 1.3 4.1c.2 1.3-.6 3.3-1.4 3.9c-1.2.8-2.6 2.3-3.6 3.3c-1.7 1.6-3.7-.3-5.6.3c-1.2.3-2.1 1.2-3.3 1.2c-1.3 0-2.5-.7-3.9-.8c-1.6-.1-3.9 1.6-5 2.5c-1.4 1.2-2.9 1.3-4.5 2c-1.5.7-2 2.1-2.4 3.7c-.4 1.5-2.3 1.6-2.7 2.4c-.9 1.5 1.3 3.5 2.3 6.9c1.2 3.8 1.7 1.6 4.8 3.3c1.2.6 1.9 2.4 3.4 3c1.7.7 3.3-.5 4.9.5c1.1.7 1.6 2.5 3 2.2c1.3-.3 2.5-2.3 3.9-1.8c3.5 1.2 3.3-.5 6.5-.4c1.7 0 3.6 1.7 5.1.3c1.4-1.2 2.3-3.1 4.1-3.3c3.5-.3 3.7-3.1 6.4-4.6c3.6-2.1 4.9-6.2 5.5-8.3c.8-2.8 1.2-5.9 3.3-9.7c1.2-2.5 0-4.8-.6-7.2"
+          />
+          <path
+            fill="#f8b100"
+            d="M21.8 42.1c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.3 1.2-1 2.4-2 3.6m10.5 4.6c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.2-1 2.5-2 3.6m-7 8c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.3 1.2-1 2.5-2 3.6m14.6-1.5c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.3 1.3-1 2.5-2 3.6m12.7-7.7c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.3 1.3-1 2.5-2 3.6m-14.6-8c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7s1.4-3.7.3-5.2c-.4 1.2-1.1 2.4-2 3.6m7.5 2.9c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.3-1 2.5-2 3.6m3.2-12c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.2-1 2.4-2 3.6m-7.4-9.3c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.2-1 2.4-2 3.6m9.7-1.6c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.2-1 2.5-2 3.6m-9.9-7.4c-1 1.1-2.1 1.9-3.2 2.3c1.5 1.1 3.6.8 4.9-.7c1.3-1.5 1.4-3.7.3-5.2c-.4 1.2-1 2.4-2 3.6m-9.5.3c-1.1 1-2.4 1.6-3.5 1.9c1.4 1.3 3.5 1.3 5 0s1.9-3.5 1-5.1c-.6 1.1-1.4 2.3-2.5 3.2m-16 30.7l-1 1l1 1.1l1-1.1zm.9 7.2l-1 1l1 1l1-1zm13.2-9.1l-1 1l1 1l1.1-1zm10 5.2l-1 1l1 1l1-1zm4.1-10.6l-1 1l1 1l1-1zm10.6-.6l-1.1 1l1.1 1l1-1zm-3.6 6.3l-1 1l1 1l1-1zm-3.6 10.7l-1 1l1 1l1.1-1zm-14.1 3.2l-1 1l1 1l1-1zm-14.5-.7l-1 1l1 1l1-1zm-6.7-4.4l-1 1l1 1l1-1zm46.3-12.1l-1 1l1 1l1-1zM25.1 10.9l-1 1l1 1l1-1zM55 22l-1 1l1 1.1l1.1-1.1zm-9.5 1.8l-1 1l1 1l1-1zM49.3 13l-1 1l1 1l1-1zm-12.6 2.8l-1 1l1 1.1l1-1.1zM39.1 5l-1 1l1 1l1-1zm-9.4.5l-1 1l1 1l1.1-1zm-5.1 40.7l-1 1l1 1l1-1zm31.5-19.1l-1.1 1l1.1 1l1-1zM46.5 8.2l-1 1l1 1l1-1z"
+          />
+        </g>
+      </svg>
+    </svg>
+  );
+}
