+++ frontend/src/Components/ShrimpFishing/SharkIcon.tsx
@@ -1,0 +1,80 @@
+export function SharkIcon() {
+  return (
+    <svg
+      width="512"
+      height="512"
+      viewBox="0 0 512 512"
+      style={{
+        height: 'auto',
+      }}
+      xmlns="http://www.w3.org/2000/svg"
+      className="h-full w-full"
+    >
+      <rect
+        width="512"
+        height="512"
+        x="0"
+        y="0"
+        rx="30"
+        fill="transparent"
+        stroke="transparent"
+        strokeWidth="0"
+        strokeOpacity="100%"
+        paintOrder="stroke"
+      ></rect>
+      <svg
+        width="256px"
+        height="256px"
+        viewBox="0 0 64 64"
+        fill="currentColor"
+        x="128"
+        y="128"
+        role="img"
+        // style="display:inline-block;vertical-align:middle"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <g fill="currentColor">
+          <path fill="#b2c1c0" d="M44.8 7c-16.1 4.3-20.2 18.5-20.2 18.5L41 28.9C39.6 22.3 44.8 7 44.8 7z" />
+          <path fill="#879191" d="m28.5 21.3l13.1-1.8s.2-1.1.4-1.9c-8.8.1-13.5 3.7-13.5 3.7" />
+          <path
+            fill="#b2c1c0"
+            d="M53.1 9.4s3 3 3.8 5.9c4.7 3.2 5.1 7.9 5.1 11c0 10.9-17.2 21.4-17.2 21.4L2 38.5c0-4.3 19.2-19 39-19c6.9 0 16 4.9 18.1 2.6c1.2-1.3-1.3-5.3-2.5-5.3c0 0 .1 2.6-1.2 4.2c.5-5.9-2.3-11.6-2.3-11.6"
+          />
+          <path fill="#ff717f" d="M19.7 55.2s0-17 16.6-23.8c0 0 9.5 19.2-16.6 23.8" />
+          <path fill="#ed4c5c" d="M25.5 38.3S22 48.4 21.9 52.8c0 0 16.5-5.1 12.5-20.6l-8.9 6.1" />
+          <path fill="#3e4347" d="M29.6 35.5s-3.5 6-4.1 11.5c-.6 5.5 8.5-2.8 9.3-13.6l-5.2 2.1" opacity=".5" />
+          <path
+            fill="#fff"
+            d="m15.9 48.2l3.3 2.1l-.6-3.9l3 2.2l-.6-4l3.5 2.6l-1.2-4.9l3.8 3.2l-1-5.3l3.9 3.2l-1.3-5.2l3.7 2.6l-1.2-4.8l3.4 3.3l-1.1-5.4s-8.4 5.3-17.6 14.3"
+          />
+          <path
+            fill="#b2c1c0"
+            d="m19.2 50.3l-3.3-2.1l1.2-1.1zm-.7-3.9l3 2.2l-2-3.8zm2.4-1.8l3.5 2.6l-2.5-4.5zm2.4-2.3l3.8 3.2l-2.5-4.6zm2.7-2.1l3.9 3.2l-2.8-5zm2.6-2l3.7 2.6l-2.7-4.2zm2.6-2.2l3.4 3.3l-2.3-4.6z"
+          />
+          <g fill="#fff">
+            <path d="m35.4 40.9l-2.2-1l1.3 3.8c.5-1.1.8-2.1.9-2.8m-1.1 3.2l-3.2-1.3l1.4 4c.8-1 1.3-1.9 1.8-2.7m-2.2 3.1l-3.2-1.8l.9 3.9c.8-.7 1.6-1.4 2.3-2.1m-2.7 2.4l-3.3-2.3l.8 3.9c1-.5 1.8-1 2.5-1.6m-2.8 1.9l-3.4-2.6s.5 3.2.6 4c1.1-.4 2-.9 2.8-1.4" />
+            <path d="m24 52.9l-3.5-2.8c.2 1.6.7 4 .7 4c1-.4 1.9-.8 2.8-1.2" />
+          </g>
+          <path
+            fill="#b2c1c0"
+            d="m35 42.4l-1.9-2.6l1.3 3.8c.3-.3.5-.8.6-1.2m-1.7 3.2l-2.2-2.9l1.4 4c.3-.3.6-.7.8-1.1m-2.5 2.9l-2-3.1l.9 3.9c.4-.2.8-.5 1.1-.8m-2.7 2l-2-3.2l.8 3.9c.5-.2.9-.4 1.2-.7m-3 1.9L23.3 49l.6 4.1c.4-.3.8-.5 1.2-.7m-2.8 1.3l-1.7-3.6l.7 4c.3-.1.6-.3 1-.4m-6.6-5.1s23.6-15.7 21.8-18.3L24 35.7l-8.3 12.9"
+          />
+          <circle cx="24.7" cy="29.6" r="3.3" fill="#3e4347" />
+          <path fill="#76807f" d="M15.9 32.1s4 .8 6-2.5s4.8-3.5 8.6-3.7c0 0-6.1-1.5-9.1.8s-2.8 4.7-5.5 5.4" />
+          <path
+            fill="#3e4347"
+            d="M44 28.6s1.4-.9 1.6-1.9c.3-1.1 1.2-3.2-.1-4.4c0 0 1.8 1 1.1 3.6c-.6 2.5-1.5 2.5-2.6 2.7m1.7.8s1.4-.7 1.8-1.8c.4-1 1.4-3.1.3-4.4c0 0 1.7 1.1.8 3.6c-.9 2.6-1.8 2.5-2.9 2.6m2.2.8s1.4-.8 1.7-1.8c.3-1 1.4-3.1.1-4.4c0 0 1.7 1.1.9 3.6c-.7 2.5-1.6 2.5-2.7 2.6"
+            opacity=".5"
+          />
+          <path fill="#879191" d="M51.3 38.3s3.9-.8 4.8 0l-1.5 1.5l-3.3-1.5" />
+          <path fill="#b2c1c0" d="M51.3 38.3C59.8 38.3 62 57 62 57s-8.1-11-18.4-11l7.7-7.7z" />
+          <path
+            fill="#e8e8e8"
+            d="M2 38.5c0-1.1 11.9-3 20.3-.2s10.6-12.4 14.8-11.2c4.6 1.3 1 15.6 10.4 19.4c-8.7 7.8-21.6 12.6-27.7 8.7c23.9-7.9 16.9-22.4 16.9-22.4S26.6 45 15.9 48.7C12 50.1 2 42 2 38.5z"
+          />
+          <path fill="#3e4347" d="M14.9 39.5s1.6-1 3 0c.6.4-.1 1.3-1.3 1.3c-1.1-.1-.7-1-1.7-1.3" />
+        </g>
+      </svg>
+    </svg>
+  );
+}
