+++ frontend/src/Components/ShrimpFishing/ShrimpIcon.tsx
@@ -1,0 +1,57 @@
+export function ShrimpIcon() {
+  return (
+    <svg
+      width="512"
+      height="512"
+      viewBox="0 0 512 512"
+      style={{
+        height: 'auto',
+      }}
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <rect
+        width="512"
+        height="512"
+        x="0"
+        y="0"
+        rx="30"
+        fill="transparent"
+        stroke="transparent"
+        strokeWidth="0"
+        strokeOpacity="100%"
+        paintOrder="stroke"
+      ></rect>
+      <svg
+        width="256px"
+        height="256px"
+        viewBox="0 0 64 64"
+        fill="currentColor"
+        x="128"
+        y="128"
+        role="img"
+        // style="display:inline-block;vertical-align:middle"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <g fill="currentColor">
+          <path fill="#ed4c5c" d="M37.7 13c-7.3-7.1-17 1.5-29.9-11c8 12.9 22.4 6.5 28.5 12.3l1.4-1.3" />
+          <path
+            fill="#ff717f"
+            d="M48.6 16.5L59 11.1c-15.9 3.7-22.7-.5-25.6 4.4c-7.4-1.2-9.8 3.5-9.8 3.5s-7.6.8-8 6.5c-2.8.3-6.3 2.9-4.9 5.7c-3.4 2-5.9 4.7-3.3 7.3c-3.7 4.3-2.3 6.2-1.4 6.6c-.9 6.6 1.5 7.2 2.3 7.4c.1 2.6.9 6.1 4 9.6c0 0 2.4-2.3.8-5.3c0 0 1.7 2.1 4.2 1.8c0 0 .2-2.3-1.3-5.1c0 0 1.9 1.7 4.2.4c0 0-2.4-3.7-6.6-5.1c.1-1.1-.2-2.5-1-4.2c.8-1.1 1.3-2.7 1.4-5.1c1.3-.9 2.5-2.6 3.3-5.5c1.1-.5 2.4-1.5 3.6-3.2c2.2.3 5.1-.5 8.5-3.5c2.2 0 5-1.1 8.3-4.2c3.5.4 9-.8 17-5.3c-.2.1-1.2.4-6.1-1.3"
+          />
+          <path
+            fill="#fff"
+            d="M43.9 15.7c-.8 2.2-3.2 3.4-5.4 2.6c-2.2-.8-3.4-3.2-2.6-5.4c.8-2.2 3.2-3.4 5.4-2.6c2.2.8 3.4 3.2 2.6 5.4"
+          />
+          <path
+            fill="#3e4347"
+            d="M41.8 15c-.4 1.1-1.5 1.6-2.6 1.3c-1.1-.4-1.6-1.5-1.3-2.6c.4-1.1 1.5-1.6 2.6-1.3c1.1.4 1.7 1.5 1.3 2.6"
+          />
+          <g fill="#ed4c5c">
+            <path d="M12.3 54.7c.5 2.6.1 7.3.1 7.3c1.9-1.3 1.6-4.2 1.6-4.2c1.4.9 3.4.8 3.4.8c-3.8-1.5-5.1-3.9-5.1-3.9" />
+            <path d="M37.1 30c.2-.4.6-.4.1-.9c-1-.9-4.8.6-8.4-1.4c.2-.1.3-.3.5-.4c2.1.1 4.9-.9 8.3-4.3c3.5.4 9-.8 16.9-5.3c-1.1.5-17.6 7.4-21.9-1c-.6 1.8-.2 3.9 1.8 5.2c-6.4 0-17.3-10.7-29.4-5.5c12.5-2.6 19.3 5.5 27 7.4c-2.3 1.1-5.5 1.8-8.5-.5c0 0 1.1 2.3 3.4 3.4c-5.2 3.7-11.5-1.4-11.5-1.4c1 2.5 2.2 4 3.5 4.8c-3.7 4.3-8.4.9-8.4.9c1.7 2 3.2 3 4.6 3.2c-.8 2.2-3.1 6.2-7.9 4.2c0 0 2.3 1.6 4.8 1.5c-.1 1.6-.8 5.5-6.2 5.1c0 0 2.8 1.5 5.1.8c.7 1.6 1.3 4.8-2.8 6.6c0 0 5.1.3 5.6-3.2c1.8 1.2 2.9 2.7 2.9 2.7c-1 .2-3.4-.2-3.4-.2c2.7 2.6 4 6.8 4 6.8c.2-2.7-1.3-5.1-1.3-5.1c2.1 1.8 4.2.4 4.2.4c-2.2-3.2-4.8-4.5-6.4-5c0-1-.3-2.3-1.1-4c.9-1 1.5-2.7 1.4-5.6c1.1-.6 2.1-1.9 2.7-4c.9 1.7 2.8 4.3 6.1 5.7c.3.1.6.3 1-.2c3.2 2.9 2.4 6 3.1 6c.4.4 1 1.3 1.1 3.2c0 0 1.2-2 1-4.2c1 0-1.2-4.4-4.4-6.6c.4-.3.7-.2.5-.8c-.6-1.3-5-1.1-7.5-4.7c1.3-.6 2.3-1.8 3-2.8c.8 1.2 3.5 4.5 8 5.4c.3.1.7.1.9-.4c3.4 2.1 3.7 5.4 4.3 5.3c.4.3 1.2 1.1 1.8 2.9c0 0 .6-2.1.1-4.3c.9-.3-2.2-4-5.7-5.6c.3-.3.7-.3.3-.8c-.8-1.1-4.6-.3-7.8-2.6c1.8-.3 3.5-1.2 4.8-2.2c1.3 1.2 4.4 3.5 8.5 3.5c.3 0 .7 0 .8-.5c4.1 1.7 4.6 4.7 5.3 4.5c.5.2 1.4.8 2.3 2.5c0 0 .3-2.5-.7-4.2c1.1-.3-2.7-3.4-6.4-4.3" />
+          </g>
+        </g>
+      </svg>
+    </svg>
+  );
+}
